[dev] separate models
